Use cache for both streamed and non-stream versions of list objects
