+++ pkg/services/authz/zanzana/server/server_list.go
@@ -2,6 +2,10 @@ package server
 
 import (
 	"context"
+	"encoding/base64"
+	"errors"
+	"hash/fnv"
+	"io"
 	"strings"
 
 	authzv1 "github.com/grafana/authlib/authz/proto/v1"
@@ -45,13 +49,6 @@ func (s *Server) List(ctx context.Context, r *authzv1.ListRequest) (*authzv1.Lis
 	}
 
 	return s.listTyped(ctx, r.GetSubject(), relation, resource, contextuals, store)
-}
-
-func (s *Server) listObjects(ctx context.Context, req *openfgav1.ListObjectsRequest) (*openfgav1.ListObjectsResponse, error) {
-	if s.cfg.UseStreamedListObjects {
-		return s.streamedListObjects(ctx, req)
-	}
-	return s.openfga.ListObjects(ctx, req)
 }
 
 func (s *Server) listTyped(ctx context.Context, subject, relation string, resource common.ResourceInfo, contextuals *openfgav1.ContextualTupleKeys, store *storeInfo) (*authzv1.ListResponse, error) {
@@ -126,6 +123,89 @@ func (s *Server) listGeneric(ctx context.Context, subject, relation string, reso
 		Folders: folderObject(folders),
 		Items:   genericObjects(resource.GroupResource(), objects),
 	}, nil
+}
+
+func (s *Server) listObjects(ctx context.Context, req *openfgav1.ListObjectsRequest) (*openfgav1.ListObjectsResponse, error) {
+	fn := s.openfga.ListObjects
+	if s.cfg.UseStreamedListObjects {
+		fn = s.streamedListObjects
+	}
+
+	if s.cfg.CheckQueryCache {
+		return s.listObjectCached(ctx, req, fn)
+	}
+
+	return fn(ctx, req)
+}
+
+type listFn func(ctx context.Context, req *openfgav1.ListObjectsRequest) (*openfgav1.ListObjectsResponse, error)
+
+func (s *Server) listObjectCached(ctx context.Context, req *openfgav1.ListObjectsRequest, fn listFn) (*openfgav1.ListObjectsResponse, error) {
+	ctx, span := tracer.Start(ctx, "server.listObjectCached")
+	defer span.End()
+
+	key, err := getRequestHash(req)
+	if err != nil {
+		return nil, err
+	}
+
+	if res, ok := s.cache.Get(key); ok {
+		return res.(*openfgav1.ListObjectsResponse), nil
+	}
+
+	res, err := fn(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	s.cache.Set(key, res, 0)
+	return res, nil
+}
+
+func (s *Server) streamedListObjects(ctx context.Context, req *openfgav1.ListObjectsRequest) (*openfgav1.ListObjectsResponse, error) {
+	ctx, span := tracer.Start(ctx, "server.streamedListObjects")
+	defer span.End()
+
+	r := &openfgav1.StreamedListObjectsRequest{
+		StoreId:              req.GetStoreId(),
+		AuthorizationModelId: req.GetAuthorizationModelId(),
+		Type:                 req.GetType(),
+		Relation:             req.GetRelation(),
+		User:                 req.GetUser(),
+		Context:              req.GetContext(),
+		ContextualTuples:     req.ContextualTuples,
+	}
+
+	stream, err := s.openfgaClient.StreamedListObjects(ctx, r)
+	if err != nil {
+		return nil, err
+	}
+
+	var objects []string
+	for {
+		res, err := stream.Recv()
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				break
+			}
+			return nil, err
+		}
+		objects = append(objects, res.GetObject())
+	}
+
+	return &openfgav1.ListObjectsResponse{
+		Objects: objects,
+	}, nil
+}
+
+func getRequestHash(req *openfgav1.ListObjectsRequest) (string, error) {
+	hash := fnv.New64a()
+	_, err := hash.Write([]byte(req.String()))
+	if err != nil {
+		return "", err
+	}
+
+	return base64.StdEncoding.EncodeToString(hash.Sum(nil)), nil
 }
 
 func typedObjects(typ string, objects []string) []string {
